added init auth service; cleaning up example stuff
