--- services/example-service-go/go.mod
@@ -1,7 +1,0 @@
-module github.com/blackflow/example-service
-
-go 1.21
-
-require (
-	github.com/gin-gonic/gin v1.9.1
-)
