--- services/example-service/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "example-service",
-  "version": "1.0.0",
-  "description": "Example backend service",
-  "main": "src/index.js",
-  "scripts": {
-    "start": "node src/index.js",
-    "dev": "nodemon src/index.js",
-    "test": "jest"
-  },
-  "dependencies": {
-    "express": "^4.18.2"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.1",
-    "jest": "^29.7.0"
-  }
-}
